Update Main file for modular purpose Add list_rental_person to ui_interface file
